requirements.txt
crewai==0.28.0  # Adjust to the latest version as needed
langchain-ollama==0.1.0  # Ensure compatibility with your Ollama setup
requests==2.28.1  # For potential future API integrations
scholarly==0.1.1  # For Google Scholar scraping
spacy
flask==3.0.0
reportlab==4.2.2
litellm==1.49.1